+++ Assets/Scripts/ColorPalette/ColorType.cs
@@ -1,0 +1,12 @@
+public enum ColorType
+{
+    Background,
+    Eclipse,
+    Number,
+    By,
+    HighScoreMessage,
+    HighScoreNumber,
+    ShameOnMessage,
+    ShameOnUsername,
+    HighScoreUsername
+}