+++ Assets/Scripts/ColorPalette/DynamicColorTMP.cs
@@ -1,0 +1,18 @@
+using TMPro;
+using UnityEngine;
+
+public class DynamicColorTMP : DynamicColor
+{
+    private TextMeshProUGUI myTextMeshProUGUI;
+    [SerializeField] private ColorType color;
+
+    private void Awake()
+    {
+        TryGetComponent(out myTextMeshProUGUI);
+    }
+
+    protected override void UpdateColor()
+    {
+        myTextMeshProUGUI.color = PaletteController.CurrentPalette.GetCustomColor(color).color;
+    }
+}