+++ CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2023-08-18
 
-## [1.5.1] - 2023-08-18
+### Added
+- New Palette System to customize colors
+
+### Changed
+- Limitation in timeouts to avoid unvoluntary bans and timeouts larger than 2 weeks
+- New TwitchChat namespace
+- Updated TwitchChat package
+
+## [1.5.1] - 2023-08-16
 
 ### Added
 - Editor extension to customize TwitchOauth and make it easier to get ready